requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
drf-yasg>=1.21.5
Pillow>=10.0.0
python-dotenv>=1.0.0
django-filter>=23.5
